pyproject.toml
[project]
name = "roku-plex-nfc"
version = "0.1.0"
description = "Tap an NFC card, play a movie or TV show on your Roku via Plex"
authors = [{ name = "Adam Guymon" }]
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "jinja2>=3.1.4",
    "plexapi>=4.18.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.30.0",
]

[project.urls]
Repository = "https://github.com/abguymon/roku-plex-nfc"
